Changes related to gcc 4.4 and gcc 5.2
-Added BOOST_NO_AUTO_PTR definition
-Added -fPIC if we're using GCC
-IGRAPH now required even if visualisation code not built (i.e. no QT)
-Added an include directory for mpfr, even if we're not using a config file
-Fixed a bunch of std::size_t/int warnings
-Added a BUILD_VISUALISATION define, which is used to strip out code in RPackageloadQT.cpp
-Apparently <: is a digraph, short for '['.

+++ PMC/PMC/PMC.cpp
@@ -78,7 +78,7 @@ namespace discreteGermGrain
 		powersQ.reserve(nVertices + 1);
 		powersP.push_back(1);
 		powersQ.push_back(1);
-		for (int i = 1; i < nVertices + 1; i++)
+		for (int i = 1; i < (int)nVertices + 1; i++)
 		{
 			powersP.push_back(powersP[i - 1] * opProbability);
 			powersQ.push_back(powersQ[i - 1] * inopProbability);
